задача выполнена!

+++ src/main/java/jm/task/core/jdbc/model/User.java
@@ -16,13 +16,13 @@ public class User {
     private String lastName;
 
     @Column
-    private Byte age;
+    private int age;
 
     public User() {
 
     }
 
-    public User(String name, String lastName, Byte age) {
+    public User(String name, String lastName, int age) {
         this.name = name;
         this.lastName = lastName;
         this.age = age;
@@ -52,11 +52,11 @@ public class User {
         this.lastName = lastName;
     }
 
-    public Byte getAge() {
+    public int getAge() {
         return age;
     }
 
-    public void setAge(Byte age) {
+    public void setAge(int age) {
         this.age = age;
     }
